LPS-60768 Move
