+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/upgrade/v0_0_3/UpgradeJournalArticles.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.journal.upgrade.v1_0_0;
+package com.liferay.journal.upgrade.v0_0_3;
 
 import com.liferay.dynamic.data.mapping.model.DDMStructure;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
